+++ src/Fruits.h
@@ -6,20 +6,23 @@
 #include "stdlib.h"
 #include <vector>
 #include "utility.h"
+#include "timer.h"
+#include "snake.h"
 
 
-class Fruits {
+class Fruit {
 public:
     Texture2D sprite;
     Area spawnArea;
     float fruitSize;
     float growing;
+    Timer* respawnTimer;
     Rectangle collisionMask;
     Vector2 position;
-    Fruits(Texture2D fruitSprite ,Area gameArea);
+    Fruit(Texture2D fruitSprite ,Area gameArea);
+    ~Fruit();
     void moveFruit();
-    void draw();
-    void update();
+    void draw(Snake& snake, int& points);
     void pulse(float, float);
 };
