--- src/loadSounds.h
@@ -1,5 +1,0 @@
-#pragma once
-
-Sound BCS = LoadSound("assets/sounds/phaseJump1.ogg");  //BCS-BorderCollisionSound
-Music IGS = LoadMusicStream("assets/soundtrack/neogauge.mp3"); //IGS-InGameSoundtrack
-Sound GameOver= LoadSound("assets/voiceOver/game_over.ogg");
